chage project structure to mvvm
