+++ app/src/main/java/kh/edu/rupp/ite/e_shopping/ui/adapter/recyclerview/AdsRecyclerAdapter.kt
@@ -6,8 +6,8 @@ import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import kh.edu.rupp.ite.e_shopping.api.model.Product
 import kh.edu.rupp.ite.e_shopping.databinding.ClothesExtraAdsItemBinding
+import kh.edu.rupp.ite.e_shopping.ui.model.Product
 import kh.edu.rupp.ite.e_shopping.ui.util.Constants.Companion.IMAGES
 
 
@@ -41,7 +41,7 @@ class AdsRecyclerAdapter : RecyclerView.Adapter<AdsRecyclerAdapter.AdsViewHolder
     override fun onBindViewHolder(holder: AdsViewHolder, position: Int) {
         val product = differ.currentList[position]
         val images = product.images
-        val image = (images!![IMAGES] as List<String>)[0]
+        val image = (images!![IMAGES] as List<*>)[0]
 
         holder.binding.apply {
             Glide.with(holder.itemView).load(image).into(imgAd)
