+++ app/src/main/java/kh/edu/rupp/ite/e_shopping/ui/adapter/recyclerview/CartRecyclerAdapter.kt
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import kh.edu.rupp.ite.e_shopping.api.model.CartProduct
+import kh.edu.rupp.ite.e_shopping.ui.model.CartProduct
 import kh.edu.rupp.ite.e_shopping.databinding.CartItemBinding
 import kh.edu.rupp.ite.e_shopping.ui.util.Constants.Companion.CART_FLAG
