+++ app/src/main/java/kh/edu/rupp/ite/e_shopping/ui/adapter/recyclerview/ClothesAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import kh.edu.rupp.ite.e_shopping.api.model.Product
+import kh.edu.rupp.ite.e_shopping.ui.model.Product
 import kh.edu.rupp.ite.e_shopping.databinding.ProductItemBinding
 import kh.edu.rupp.ite.e_shopping.ui.util.Constants.Companion.IMAGES
