+++ app/src/main/java/kh/edu/rupp/ite/e_shopping/ui/model/Order.kt
@@ -1,4 +1,4 @@
-package kh.edu.rupp.ite.e_shopping.api.model
+package kh.edu.rupp.ite.e_shopping.ui.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
