+++ app/src/main/res/layout/fragment_cart.xml
@@ -5,7 +5,7 @@
                                        android:layout_width="match_parent"
                                        android:layout_height="match_parent"
                                        android:fillViewport="true"
-                                       tools:context=".ui.view.fragment.CartFragment">
+                                       tools:context=".ui.view.fragment.home.CartFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
