+++ app/src/main/res/layout/fragment_home.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:fillViewport="true"
-        tools:context=".ui.view.fragment.HomeFragment">
+        tools:context=".ui.view.fragment.home.HomeFragment">
 
     <TextView
         android:id="@+id/tv_search"
