+++ app/src/main/res/layout/fragment_home_products_view_holder.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".ui.view.fragment.HomeFragment">
+    tools:context=".ui.view.fragment.home.HomeFragment">
 
 
     <androidx.core.widget.NestedScrollView
