+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".ui.view.fragment.ProfileFragment"
+    tools:context=".ui.view.fragment.home.ProfileFragment"
     android:background="#F4F4F4">
 
 
@@ -53,7 +53,7 @@
             android:layout_width="@dimen/_43sdp"
             android:layout_height="@dimen/_43sdp"
             android:background="@drawable/profile_img_background"
-            android:src="@drawable/test_image"
+            android:src="@drawable/boy"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
