+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.fragment.SearchFragment"
+    tools:context=".ui.view.fragment.home.SearchFragment"
     android:fillViewport="true">
 
     <androidx.constraintlayout.widget.ConstraintLayout
