+++ app/src/main/res/navigation/bottom_nav.xml
@@ -4,14 +4,14 @@
             xmlns:tools="http://schemas.android.com/tools" android:id="@+id/bottom_nav"
             app:startDestination="@id/home">
 
-    <fragment android:id="@+id/home" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.HomeFragment" android:label="fragment_home"
+    <fragment android:id="@+id/home" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.home.HomeFragment" android:label="fragment_home"
               tools:layout="@layout/fragment_home">
         <action android:id="@+id/action_home_to_search" app:destination="@id/search"/>
     </fragment>
-    <fragment android:id="@+id/search" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.SearchFragment" android:label="fragment_search"
+    <fragment android:id="@+id/search" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.home.SearchFragment" android:label="fragment_search"
               tools:layout="@layout/fragment_search"/>
-    <fragment android:id="@+id/cart" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.CartFragment" android:label="fragment_cart"
+    <fragment android:id="@+id/cart" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.home.CartFragment" android:label="fragment_cart"
               tools:layout="@layout/fragment_cart"/>
-    <fragment android:id="@+id/profile" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.ProfileFragment"
+    <fragment android:id="@+id/profile" android:name="kh.edu.rupp.ite.e_shopping.ui.view.fragment.home.ProfileFragment"
               android:label="fragment_profile" tools:layout="@layout/fragment_profile"/>
 </navigation>
